functions/update_fw/requirements.txt
# Function dependencies, for example:
# package>=version

google-api-python-client==1.10.0
google-cloud-error-reporting
oauth2client
dnspython